Refactor JobStatusManager component
The JobStatusManager component was refactored into smaller, more focused components to improve maintainability.
